+++ src/config.ts
@@ -1,3 +1,3 @@
-export const CORE_API_BASE_URL = ""
-export const MOCK_API_BASE_URL =
+export const CORE_API_BASE_URL =
   "https://jndqonikai7v45mkeohsogcvqe0tzdue.lambda-url.ap-southeast-2.on.aws"
+export const API_VERSION = "v1"
